// src/pages/weather/widgets/DailyForecast/types.ts
/*
 * Copyright © 2022 Sinclair Broadcast Group
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 * 
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.

 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

interface IHourlyForecastData {
  temperature: number;
  time: string;
  iconCode?: number;
  iconName?: string;
}

interface IDailyForecastData {
  currentTemp: number;
  highestTemp: number;
  hourly: Array<IHourlyForecastData> | null;
  humidity: number;
  lowestTemp: number;
  precipitation: number;
  ultraviolet: number;
  windDirection: number;
  windSpeed: number;
  iconCode?: number;
  iconName?: string;
}

interface IDailyForecastWidgetProps {
  data: IDailyForecastData;
  loading: boolean;
}

export type { IDailyForecastWidgetProps, IDailyForecastData, IHourlyForecastData };